app/requirements.txt
Django==3.1.5
gunicorn==20.0.4
django-modelcluster
psycopg2
djangorestframework
markdown
django-filter
